+++ example/lib/home.dart
@@ -1,3 +1,4 @@
+import 'package:deeplink_x_example/pages/air_navigation_pro_page.dart';
 import 'package:deeplink_x_example/pages/apple_maps_page.dart';
 import 'package:deeplink_x_example/pages/cafe_bazaar_page.dart';
 import 'package:deeplink_x_example/pages/facebook_page.dart';
@@ -50,6 +51,11 @@ final _apps = <_DemoItem>[
   _DemoItem(title: 'Waze', assetName: 'assets/waze.png', builder: (_) => const WazePage()),
   _DemoItem(title: 'Apple Maps', assetName: 'assets/apple_maps.png', builder: (_) => const AppleMapsPage()),
   _DemoItem(title: 'Sygic', assetName: 'assets/sygic.png', builder: (_) => const SygicPage()),
+  _DemoItem(
+    title: 'Air Navigation Pro',
+    assetName: 'assets/air_navigation_pro.png',
+    builder: (_) => const AirNavigationProPage(),
+  ),
 ];
 
 // List of store demos.